requirements.txt
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
et_xmlfile==2.0.0
h11==0.16.0
idna==3.11
numpy==2.3.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
selenium==4.37.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
